requirements.txt
pillow==11.3.0
facebook-sdk==3.1.0
dotenv==0.9.9
requests==2.32.4
